(WIP) Add friend operator to print state and debug.

+++ util/pd/inc/dp.h
@@ -3,6 +3,6 @@
 
 #include "util.h"
 
-int dp(struct Game&, struct State&);
+short dp(struct Game&, struct State&);
 
 #endif
